dashboard/requirements.txt
python-dotenv
psycopg2
streamlit
st-annotated-text